requirements.txt
python-dotenv
google-cloud-firestore
google-cloud-secret-manager
sqlalchemy
pyodbc
sqlglot
polars
google-cloud-storage
